Releasing

+++ src/TD.Web/TD.Web.Admin/TD.Web.Admin.Contracts/Dtos/ProductsPrices/ProductsPricesGetDto.cs
@@ -2,8 +2,8 @@
 {
     public class ProductsPricesGetDto
     {
-        public int Id {  get; set; }
-        public int ProductId { get; set; }
+        public long Id {  get; set; }
+        public long ProductId { get; set; }
         public decimal Min { get; set; }
         public decimal Max { get; set; }
     }
